# Modify redirect_to to halt further execution.