This PR is a work in progress as there are still some tests that are not passing. I am submitting the PR in hope to receive some feedback about this change and what other additions would be necessary to get it in. I realize this is a breaking change, and perhaps this change could be introduced in a similar route as #17227 where the continuing of execution is only deprecated until a version increase. Alternatively, I spoke with @tenderlove at RailsConf earlier today who brought up the concern about potential failures with transactions as this exception should _not_ result in transaction failures. I'm not sure the best way to handle that concern. In our very brief conversation, @tenderlove also was more comfortable with the separate notion of introducing a `redirect_to!` method that results in the halting of action code. However, my prior (dated) [research](http://adamdoupe.com/publications/fear-the-ear-ccs2011.pdf) suggests that the majority of uses of `redirect_to` do not require execution after redirect. Hence I would prefer to not add an additional method.

Assuming this works as planned, I think similar changes could be made to `head` and `render` as they also seldom require execution after their respective calls.

---

The purpose of this addition is to prevent possible execution after redirect exceptions in rails. While the suggested method of handling control flow is to utilize a `before_filter` to prevent code in a controller from executing, there are cases were rails developers will legitimately use `redirect_to` in their controllers.

With that in mind, there is zero reason for any code, save for ensure blocks, to execute after a `redirect_to` is issued. Anywhere someone intentionally executes code following a redirect, it can be restructured to occur prior to the `redirect_to` exception. If modifying the redirect response body is desired feature, then `redirect_to` should be extended to take a custom response body as a parameter.

This is a reimplementation of my friend and colleague Adam Doupé's prior work:
adamdoupe@78c6550

---

Updated to use throw/catch rather than raise/rescue. Previous commit:
e4953d3