@kaspth 2015-04-24
Can you try benchmarking this and see if there's any impact by this change? You can use https://github.com/eileencodes/integration_performance_test.

Also please take the company name out of the commit message. We're not here to promote it. Thanks ❤️