@bboe 2015-04-24
I removed the company acknowledgement. Yes, would love to benchmark the change. Could someone please offer a link or a quick suggestion on how using rvm can test my rails branch?