@bboe 2015-04-26
I ran the tool suggested by @kaspth. The results between subsequent runs of the tool regardless of branch were inconsistent in both directions, thus I don't think a direct comparison between these two branches is accurate (I certainly wouldn't use these results in any publication).

On master:

```sh
$ ruby -Itest test/controllers/documents_benchmark_test.rb --seed 1
DEPRECATION WARNING: alias_method_chain is deprecated. Please, use Module#prepend instead. From module, you can access the original method using super. (called from included at /Users/bboe/.rvm/gems/ruby-2.2.2/bundler/gems/turbolinks-81404b03f36a/lib/turbolinks/xhr_url_for.rb:7)
Calculating -------------------------------------
INDEX: Integration Test
                        93.000  i/100ms
INDEX: Functional Test
                       112.000  i/100ms
-------------------------------------------------
INDEX: Integration Test
                        971.995  (± 4.1%) i/s -      4.929k
INDEX: Functional Test
                          1.117k (± 3.8%) i/s -      5.600k

Comparison:
INDEX: Functional Test:     1116.7 i/s
INDEX: Integration Test:      972.0 i/s - 1.15x slower

Calculating -------------------------------------
CREATE: Integration Test
                        34.000  i/100ms
CREATE: Functional Test
                        42.000  i/100ms
-------------------------------------------------
CREATE: Integration Test
                        344.118  (± 7.6%) i/s -      1.734k
CREATE: Functional Test
                        417.475  (± 5.7%) i/s -      2.100k

Comparison:
CREATE: Functional Test:      417.5 i/s
CREATE: Integration Test:      344.1 i/s - 1.21x slower

Run options: --seed 1

# Running:

....

Finished in 0.012688s, 315.2585 runs/s, 472.8878 assertions/s.

4 runs, 6 assertions, 0 failures, 0 errors, 0 skips
```

with this PR:

```sh
$ ruby -Itest test/controllers/documents_benchmark_test.rb --seed 1
DEPRECATION WARNING: alias_method_chain is deprecated. Please, use Module#prepend instead. From module, you can access the original method using super. (called from included at /Users/bboe/.rvm/gems/ruby-2.2.2/bundler/gems/turbolinks-81404b03f36a/lib/turbolinks/xhr_url_for.rb:7)
Calculating -------------------------------------
INDEX: Integration Test
                        95.000  i/100ms
INDEX: Functional Test
                       111.000  i/100ms
-------------------------------------------------
INDEX: Integration Test
                        968.694  (± 5.5%) i/s -      4.845k
INDEX: Functional Test
                        998.176  (±13.2%) i/s -      4.995k

Comparison:
INDEX: Functional Test:      998.2 i/s
INDEX: Integration Test:      968.7 i/s - 1.03x slower

Calculating -------------------------------------
CREATE: Integration Test
                        28.000  i/100ms
CREATE: Functional Test
                        31.000  i/100ms
-------------------------------------------------
CREATE: Integration Test
                        338.849  (± 6.8%) i/s -      1.708k
CREATE: Functional Test
                        417.339  (± 7.4%) i/s -      2.077k

Comparison:
CREATE: Functional Test:      417.3 i/s
CREATE: Integration Test:      338.8 i/s - 1.23x slower

Run options: --seed 1

# Running:

....

Finished in 0.010944s, 365.4971 runs/s, 548.2456 assertions/s.

4 runs, 6 assertions, 0 failures, 0 errors, 0 skips
```