@maclover7 2016-01-10
Thanks for submitting this PR! Unfortunately, I don't really understand what is trying to be accomplished here. To me, there's no real difference between the [first]() and [second](https://github.com/rails/rails/pull/19883/files#diff-e3e7233ead710d24a9c88db3fd7e61baR123) tests. If people want to raise errors, it should be convention to put it above the redirect_to. On a related note, you benchmarks show a decrease in performance by 2%. Thank you again for your contribution.